requirements.txt
pip
autopep8
cloudinary==1.38.0
Deprecated==1.2.14
email_validator==2.2.0
Flask==3.0.1
Flask-Admin==1.6.1
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
gunicorn==21.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
openpyxl==3.1.5
pandas==2.2.2
phonenumbers==8.13.45
psycopg2==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
pytest-flask==1.3.0
python-dotenv==1.0.1
python-slugify==8.0.3
requests==2.32.3
SQLAlchemy==2.0.31
text-unidecode==1.3
typing_extensions==4.12.2
Werkzeug==3.0.3